% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{archive_invoice}
\alias{archive_invoice}
\title{Archive Invoice}
\usage{
archive_invoice(authorization, invoice_id)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Archive Invoice
}
