% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{check_authorization}
\alias{check_authorization}
\title{Check Authorization}
\usage{
check_authorization(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{amount}{int32. REQUIRED
Amount in kobo}

\item{authorization_code}{string. REQUIRED
Authorization code representing card you want charged}

\item{email}{string. REQUIRED
Customer's email address}

\item{currency}{string.
A currency for the amount we want to check}
}
\value{

}
\description{
Check Authorization
}
\details{
In test mode, we will return insufficient funds for an amount greater than or equal 500,000 naira.
This endpoint will only work for authorizations with for which brand is mastercard or visa
}
