% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{check_balance}
\alias{check_balance}
\title{Check Balance}
\usage{
check_balance(authorization)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}
}
\value{

}
\description{
Check Balance
}
