% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{create_customer}
\alias{create_customer}
\title{Create Customer}
\usage{
create_customer(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Create Customer
}
