% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{create_invoice}
\alias{create_invoice}
\title{Create Invoice}
\usage{
create_invoice(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Create Invoice
}
