% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{create_plan}
\alias{create_plan}
\title{Create Plan}
\usage{
create_plan(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Create Plan
}
