% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refund.R
\name{create_refund}
\alias{create_refund}
\title{Create Refund}
\usage{
create_refund(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{transaction}{string. REQUIRED
Transaction reference or id}

\item{amount}{int32.
How much in kobo to be refunded to the customer. Amount is optional(defaults to original transaction amount) and cannot be more than the original transaction amount.}

\item{currency}{string. Three-letter ISO currency}

\item{customer_note}{string. Customer reason}

\item{merchant_note}{string. Merchant reason}
}
\value{

}
\description{
This creates a refund which is then processed by the Paystack team
}
