% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subaccount.R
\name{create_subaccount}
\alias{create_subaccount}
\title{Create Subaccount}
\usage{
create_subaccount(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{account_number}{string. REQUIRED
NUBAN Bank Account Number}

\item{business_name}{string. REQUIRED
Name of business for subaccount}

\item{percentage_charge}{float. REQUIRED
What is the default percentage charged when receiving on behalf of this subaccount?}

\item{settlement_bank}{string. REQUIRED
Name of Bank (see list of accepted names by calling List Banks)}

\item{primary_contact_email}{string.
A contact email for the subaccount}

\item{primary_contact_name}{string.
A name for the contact person for this subaccount}

\item{primary_contact_phone}{string.
A phone number to call for this subaccount}

\item{metadata}{string.
Stringified JSON object. Sample: {"custom_fields":[{"display_name":"Cart ID","variable_name":"cart_id","value":"8393"}]}}

\item{settlement_schedule}{string.
Any of auto, weekly, monthly, manual. Auto means payout is T+1 and manual means payout to the subaccount should only be made when requested.}
}
\value{

}
\description{
Create Subaccount
}
