% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{delete_transfer_recipients}
\alias{delete_transfer_recipients}
\title{Delete Transfer Recipients}
\usage{
delete_transfer_recipients(authorization, recipient_id)
}
\arguments{
\item{recipient_id}{}
}
\value{

}
\description{
Delete Transfer Recipients
}
