% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{disable_otp}
\alias{disable_otp}
\title{Disable OTP requirement for Transfer}
\usage{
disable_otp(authorization)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}
}
\value{

}
\description{
In the event that you want to be able to complete transfers programmatically without use of OTPs, this endpoint helps disable that….with an OTP. No arguments required. You will get an OTP.
}
\details{
In the event that you want to be able to complete transfers programmatically without use of OTPs, this endpoint helps disable that….with an OTP. No arguments required. An OTP is sent to you on your business phone.
}
