% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{disable_subscription}
\alias{disable_subscription}
\title{Disable Subscription}
\usage{
disable_subscription(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{code}{string. REQUIRED
Subscription code}

\item{token}{string. REQUIRED
Email token}
}
\value{

}
\description{
Disable Subscription
}
