% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{enable_otp}
\alias{enable_otp}
\title{Enable OTP requirement for Transfer}
\usage{
enable_otp(authorization)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}
}
\value{

}
\description{
Enable OTP requirement for Transfer
}
