% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{export_transactions}
\alias{export_transactions}
\title{Export Transactions}
\usage{
export_transactions(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Export Transactions
}
