% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkcharge.R
\name{fetch_bulk_charge_batch}
\alias{fetch_bulk_charge_batch}
\title{Fetch Bulk Charge Batch}
\usage{
fetch_bulk_charge_batch(authorization, batch_id)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{batch_id}{}
}
\value{

}
\description{
Fetch Bulk Charge Batch
}
