% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{fetch_customer}
\alias{fetch_customer}
\title{Fetch Customer}
\usage{
fetch_customer(authorization, id)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{id}{int32. REQUIRED Path Param.
An email or ID or customer code for the customer you want to fetch}
}
\value{

}
\description{
Fetch Customer
}
