% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{fetch_page}
\alias{fetch_page}
\title{Fetch Page}
\usage{
fetch_page(authorization, page_id)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{page_id}{or slug string. REQUIRED Path param}
}
\value{

}
\description{
Fetch Page
}
