% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{fetch_plan}
\alias{fetch_plan}
\title{Fetch Plan}
\usage{
fetch_plan(authorization, plan_id)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{plan_id}{}
}
\value{

}
\description{
Fetch Plan
}
