% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refund.R
\name{fetch_refund}
\alias{fetch_refund}
\title{Fetch Refund}
\usage{
fetch_refund(authorization, trans_id)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{trans_id}{}
}
\value{

}
\description{
Fetch Refund
}
