% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settlement.R
\name{fetch_settlements}
\alias{fetch_settlements}
\title{Fetch Settlements}
\usage{
fetch_settlements(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Fetch Settlements
}
