% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subaccount.R
\name{fetch_subaccount}
\alias{fetch_subaccount}
\title{Fetch Subaccount}
\usage{
fetch_subaccount(authorization, subacct_id)
}
\arguments{
\item{subacct_id}{}
}
\value{

}
\description{
Fetch Subaccount
}
