% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{fetch_transfer}
\alias{fetch_transfer}
\title{Fetch Transfers}
\usage{
fetch_transfer(authorization, ...)

fetch_transfer(authorization, ...)
}
\arguments{
\item{...}{}

\item{recipient_id}{}
}
\value{



}
\description{
Fetch Transfers

Finalize Transfers
}
