% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{finalize_disable_otp}
\alias{finalize_disable_otp}
\title{Finalize Disabling OTP requirement for Transfer}
\usage{
finalize_disable_otp(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Finalize Disabling OTP requirement for Transfer
}
