% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{finalize_invoice}
\alias{finalize_invoice}
\title{Finalize a Draft Invoice}
\usage{
finalize_invoice(authorization, invoice_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{send_notification}{boolean.
Indicates whether Paystack sends an email notification to customer. Defaults to true}
}
\value{

}
\description{
Finalize a Draft Invoice
}
