% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{finalize_invoices}
\alias{finalize_invoices}
\title{Finalize a Draft Invoice}
\usage{
finalize_invoices(authorization, invoice_id, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Finalize a Draft Invoice
}
