% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{get_invoice_metrics}
\alias{get_invoice_metrics}
\title{Get Invoice Metrics}
\usage{
get_invoice_metrics(authorization)
}
\arguments{
\item{authorization}{}
}
\value{

}
\description{
Get Invoice Metrics
}
