% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{initialize_transaction}
\alias{initialize_transaction}
\title{Initialize Transaction}
\usage{
initialize_transaction(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Initialize Transaction
}
