% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{initiate_transfer}
\alias{initiate_transfer}
\title{Initiate Transfer}
\usage{
initiate_transfer(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Initiate Transfer
}
