% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ids.R
\name{is_id_available}
\alias{is_id_available}
\title{Check Available}
\usage{
is_id_available(id, type, ...)
}
\arguments{
\item{id}{to search for}

\item{type}{category to check}

\item{...}{authorization}
}
\value{
boolean
}
\description{
Check if an id is available for a particular section of account, type
}
\examples{
is_id_available("se546", "transaction", authorization)
}
