% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\name{list_banks}
\alias{list_banks}
\title{List Banks}
\usage{
list_banks(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
List Banks
}
