% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkcharge.R
\name{list_bulk_charge_batches}
\alias{list_bulk_charge_batches}
\title{List Bulk Charge Batches}
\usage{
list_bulk_charge_batches(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
List Bulk Charge Batches
}
