% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{list_pages}
\alias{list_pages}
\title{List Pages}
\usage{
list_pages(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{perPage}{int32.
Specify how many records you want to retrieve per page}

\item{page}{int32.
Specify exactly what page you want to retrieve}
}
\value{

}
\description{
List Pages
}
