% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{list_plans}
\alias{list_plans}
\title{List Plans}
\usage{
list_plans(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{perPage}{int32.
Specify how many records you want to retrieve per page}

\item{page}{int32.
Specify exactly what page you want to retrieve}

\item{interval}{string.
Filter list by plans with specified interval}

\item{amount}{string.
Filter list by plans with specified amount (in kobo)}
}
\value{

}
\description{
List Plans
}
