% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refund.R
\name{list_refunds}
\alias{list_refunds}
\title{List Refunds}
\usage{
list_refunds(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
List Refunds
}
