% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subaccount.R
\name{list_subaccount}
\alias{list_subaccount}
\title{List Subaccounts}
\usage{
list_subaccount(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
List Subaccounts
}
