% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{list_subscriptions}
\alias{list_subscriptions}
\title{List Subscriptions}
\usage{
list_subscriptions(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{perPage}{int32.
Specify how many records you want to retrieve per page}

\item{page}{int32.
Specify exactly what page you want to retrieve}

\item{customer}{int32.
Filter by Customer ID}

\item{plan}{int32.
Filter by Plan ID}
}
\value{

}
\description{
List Subscriptions
}
