% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{list_transfer_recipients}
\alias{list_transfer_recipients}
\title{List Transfer Recipients}
\usage{
list_transfer_recipients(authorization, ...)

list_transfer_recipients(authorization, ...)
}
\arguments{
\item{...}{}

\item{...}{}
}
\value{



}
\description{
List Transfer Recipients

List Transfers
}
