% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{mark_as_paid}
\alias{mark_as_paid}
\title{Mark as Paid}
\usage{
mark_as_paid(authorization, invoice_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{amount_paid}{int32. REQUIRED
Amount paid by customer}

\item{paid_by}{string. REQUIRED
The email address of the customer}

\item{payment_date}{string. REQUIRED
Date payment was made}

\item{payment_method}{string. REQUIRED
The method of payment which are "Cash", "POS", "Cheque", "Bank Transfer" or "Other"}

\item{note}{string.
An optional note}
}
\value{

}
\description{
This marks an invoice as paid in cases where it was paid offline
}
