% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_response}
\alias{parse_response}
\title{Parse json response to list}
\usage{
parse_response(response)
}
\arguments{
\item{response}{A response object}
}
\value{

}
\description{
Takes in an object with a class of "response" returned from calling any of the wrapper functions.
Also pipe friendly.
}
\details{
As long as "status": true,
returned item would most likely be a deep list,
which means apart from the status and message,
other elements of the returned lists would be lists as well
}
\examples{
bank_list <- list_banks(auth$secret)
parse_response(bank_list)
}
