% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkcharge.R
\name{pause_bulk_charge_batch}
\alias{pause_bulk_charge_batch}
\title{Pause Bulk Charge Batch}
\usage{
pause_bulk_charge_batch(authorization, batch_code)
}
\arguments{
\item{batch_code}{}
}
\value{

}
\description{
Pause Bulk Charge Batch
}
