% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{resend_otp}
\alias{resend_otp}
\title{Resend OTP for Transfer}
\usage{
resend_otp(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Resend OTP for Transfer
}
