% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\name{resolve_account_number}
\alias{resolve_account_number}
\title{Resolve Account Number}
\usage{
resolve_account_number(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{account_number}{string. Account Number}

\item{bank_code}{string. Bank Code}
}
\value{

}
\description{
Resolve Account Number
}
