% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\name{resolve_bvn}
\alias{resolve_bvn}
\title{Resolve BVN}
\usage{
resolve_bvn(authorization, ...)

resolve_bvn(authorization, ...)
}
\arguments{
\item{...}{}

\item{bvn}{}
}
\value{



}
\description{
Resolve BVN

Resolve Account Number
}
