% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_card.R
\name{resolve_card_bin}
\alias{resolve_card_bin}
\title{Resolve Card BIN}
\usage{
resolve_card_bin(card_bin)
}
\arguments{
\item{card_bin}{string. REQUIRED
First 6 characters of card}
}
\value{

}
\description{
No authorization required
}
