% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_phone.R
\name{resolve_phone_number}
\alias{resolve_phone_number}
\title{Resolve Phone Number}
\usage{
resolve_phone_number(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{verification_type}{string. REQUIRED}

\item{phone}{string. REQUIRED
Customer phone number starting with country code (without the + prefix)}

\item{callback_url}{string. REQUIRED
Link on server to receive the truecaller details}
}
\value{

}
\description{
Using the Truecaller API you can verify the authenticity of a customer. It returns the customer's name, phone number, email, social media handles and organization as available on their Truecaller profile.
}
