% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{send_notification}
\alias{send_notification}
\title{Send Notification for Payment request}
\usage{
send_notification(authorization, invoice_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Send Notification for Payment request
}
