% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_keys.R
\name{set_keys}
\alias{set_keys}
\title{Set API Keys}
\usage{
set_keys(public_key, secret_key)
}
\arguments{
\item{public_key}{Public Key}

\item{secret_key}{Secret Key}
}
\value{
equivalent of -H "Authorization: Bearer SECRET_KEY"
}
\description{
Call this function to store API keys in an R object either for test or live modes
}
\examples{
test_keys <- set_keys("pk_test_key", "sk_test_key")
live_keys <- set_keys("pk_live_key", "sk_live_key")

}
