% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction_totals}
\alias{transaction_totals}
\title{Transaction Totals}
\usage{
transaction_totals(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Transaction Totals
}
