% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{update_customer}
\alias{update_customer}
\title{Update Customer}
\usage{
update_customer(authorization, id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{id}{string. REQUIRED Path Param.
Unique customer_id to update}

\item{...}{Body Params}

\item{first_name}{string.
Customer's first name}

\item{last_name}{string.
Customer's last name}

\item{phone}{string.
Customer's phone number}

\item{metadata}{object.
A set of key/value pairs that you can attach to the customer. It can be used to store additional information in a structured format.}
}
\value{

}
\description{
Update Customer
}
