% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{update_invoice}
\alias{update_invoice}
\title{Update Invoice}
\usage{
update_invoice(authorization, invoice_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Update Invoice
}
