% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{update_page}
\alias{update_page}
\title{Update Page}
\usage{
update_page(authorization, page_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{name}{string. REQUIRED
Name of page}

\item{description}{string.
Short description of page}

\item{amount}{int32.
Default amount you want to accept using this page. If none is set, customer is free to provide any amount of their choice. The latter scenario is useful for accepting donations}

\item{active}{boolean.
Set to false to deactivate page url}
}
\value{

}
\description{
Update Page
}
