% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payment_sessions.R
\name{update_payment_session_timeout}
\alias{update_payment_session_timeout}
\title{Update Payment Session Timeout}
\usage{
update_payment_session_timeout(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{timeout}{int32.
Time before stopping session (in seconds). Set to 0 to cancel session timeouts}
}
\value{

}
\description{
Update Payment Session Timeout
}
