% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{update_plan}
\alias{update_plan}
\title{Update Plan}
\usage{
update_plan(authorization, plan_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Update Plan
}
