% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{update_transfer_recipients}
\alias{update_transfer_recipients}
\title{Update Transfer Recipients}
\usage{
update_transfer_recipients(authorization, recipient_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Update Transfer Recipients
}
