% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{verify_transaction}
\alias{verify_transaction}
\title{Verify Transaction}
\usage{
verify_transaction(authorization, reference)
}
\arguments{
\item{reference}{}
}
\value{

}
\description{
Verify Transaction
}
