% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{view_invoice}
\alias{view_invoice}
\title{View Invoice}
\usage{
view_invoice(authorization, invoice_id)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{invoice_id}{}
}
\value{

}
\description{
View Invoice
}
