% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{view_transaction_timeline}
\alias{view_transaction_timeline}
\title{View Transaction Timeline}
\usage{
view_transaction_timeline(authorization, id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
View Transaction Timeline
}
