% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{archive_invoice}
\alias{archive_invoice}
\title{Archive Invoice}
\usage{
archive_invoice(authorization, invoice_id)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Used to archive an invoice. Invoice will no longer be fetched on list or returned on verify.
}
