% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{charge_authorization}
\alias{charge_authorization}
\title{Charge Authorization}
\usage{
charge_authorization(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Charge Authorization
}
