% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{check_authorization}
\alias{check_authorization}
\title{Check Authorization}
\usage{
check_authorization(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Check Authorization
}
