% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{check_authorization}
\alias{check_authorization}
\title{Check Authorization}
\usage{
check_authorization(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Check Authorization
}
