% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge.R
\name{check_pending_charge}
\alias{check_pending_charge}
\title{Check Pending Charge}
\usage{
check_pending_charge(authorization, reference)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{reference}{}
}
\value{

}
\description{
Check Pending Charge
}
