% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{check_slug_availability}
\alias{check_slug_availability}
\title{Check Slug Availability}
\usage{
check_slug_availability(authorization, slug)
}
\arguments{
\item{slug}{}
}
\value{

}
\description{
Check Slug Availability
}
