% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{create_page}
\alias{create_page}
\title{Create Page}
\usage{
create_page(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Create Page
}
