% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{create_plan}
\alias{create_plan}
\title{Create Plan}
\usage{
create_plan(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{amount}{int32. REQUIRED
Amount to be charged in kobo}

\item{interval}{string. REQUIRED
Interval in words. Valid intervals are hourly, daily, weekly, monthly,biannually, annually.}

\item{name}{string. REQUIRED
Name of plan}

\item{description}{string.
Short description of plan}

\item{send_invoices}{boolean.
Set to false if you don't want invoices to be sent to your customers}

\item{send_sms}{boolean.
Set to false if you don't want text messages to be sent to your customers}

\item{currency}{string.
Currency in which amount is set}

\item{invoice_limit}{int32.
Number of invoices to raise during subscription to this plan. Can be overridden by specifying an invoice_limit while subscribing.}
}
\value{

}
\description{
Create Plan
}
