% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refund.R
\name{create_refund}
\alias{create_refund}
\title{Create Refund}
\usage{
create_refund(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Create Refund
}
