% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subaccount.R
\name{create_subaccount}
\alias{create_subaccount}
\title{Create Subaccount}
\usage{
create_subaccount(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Create Subaccount
}
