% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{create_subscription}
\alias{create_subscription}
\title{Create Subscription}
\usage{
create_subscription(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{authorization}{string. REQUIRED
If customer has multiple authorizations, you can set the desired authorization you wish to use for this subscription here. If this is not supplied, the customer's most recent authorization would be used}

\item{customer}{string. REQUIRED
Customer's email address or customer code}

\item{plan}{string. REQUIRED
Plan code}

\item{start_date}{string.
Set the date for the first debit. (ISO 8601 format)}
}
\value{

}
\description{
Create Subscription
}
