% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{create_subscription}
\alias{create_subscription}
\title{Create Subscription}
\usage{
create_subscription(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Create Subscription
}
