% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{create_transfer_recipient}
\alias{create_transfer_recipient}
\title{Create Transfer Recipient}
\usage{
create_transfer_recipient(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Create Transfer Recipient
}
