% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{create_transfer_recipient}
\alias{create_transfer_recipient}
\title{Create Transfer Recipient}
\usage{
create_transfer_recipient(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Create Transfer Recipient
}
