% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{deactivate_authorization}
\alias{deactivate_authorization}
\title{Deactivate Authorization}
\usage{
deactivate_authorization(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Deactivate Authorization
}
