% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{delete_transfer_recipients}
\alias{delete_transfer_recipients}
\title{Delete Transfer Recipients}
\usage{
delete_transfer_recipients(authorization, recipient_id)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{recipient_id}{string. REQUIRED
Code or ID for transfer recipient we want to update.}
}
\value{

}
\description{
Delete Transfer Recipients
}
