% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{disable_otp}
\alias{disable_otp}
\title{Disable OTP requirement for Transfer}
\usage{
disable_otp(authorization)
}
\arguments{
\item{authorization}{}
}
\value{

}
\description{
Disable OTP requirement for Transfer
}
