% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{disable_subscription}
\alias{disable_subscription}
\title{Disable Subscription}
\usage{
disable_subscription(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Disable Subscription
}
