% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{enable_subscription}
\alias{enable_subscription}
\title{Enable Subscription}
\usage{
enable_subscription(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Enable Subscription
}
