% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkcharge.R
\name{fetch_charges_in_batch}
\alias{fetch_charges_in_batch}
\title{Fetch Charges in Batch}
\usage{
fetch_charges_in_batch(authorization, batch_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{batch_id}{string. REQUIRED
An ID or code for the batch whose details you want to retrieve.}

\item{...}{Body Params}

\item{status}{string. pending, success or failed}

\item{perPage}{int32.}

\item{page}{int32.}
}
\value{

}
\description{
This endpoint retrieves the charges associated with a specified batch code. Pagination parameters are available. You can also filter by status. Charge statuses can be pending, success or failed.
}
