% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{fetch_customer}
\alias{fetch_customer}
\title{Fetch Customer}
\usage{
fetch_customer(authorization, customer_code, ...)

fetch_customer(authorization, customer_code, ...)
}
\arguments{
\item{...}{}

\item{id}{email or id or customer_code}
}
\value{



}
\description{
Fetch Customer

Update Customer
}
