% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{fetch_customer}
\alias{fetch_customer}
\title{Fetch Customer}
\usage{
fetch_customer(authorization, customer_code, ...)

fetch_customer(authorization, customer_code, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{id}{int32. REQUIRED Path Param.
An email or ID or customer code for the customer you want to fetch}

\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}
}
\value{



}
\description{
Fetch Customer

Fetch Customer
}
