% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payment_sessions.R
\name{fetch_payment_session_timeout}
\alias{fetch_payment_session_timeout}
\title{Fetch Payment Session Timeout}
\usage{
fetch_payment_session_timeout(authorization)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}
}
\value{

}
\description{
Fetch Payment Session Timeout
}
