% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settlement.R
\name{fetch_settlements}
\alias{fetch_settlements}
\title{Fetch Settlements}
\usage{
fetch_settlements(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{from}{date.
Lower bound of date range. Leave undefined to export settlement from day one.}

\item{to}{date.
Upper bound of date range. Leave undefined to export settlements till date.}

\item{subaccount}{string.
Provide a subaccount code to export only settlements for that subaccount. Set to none to export only transactions for the account.}
}
\value{

}
\description{
Fetch Settlements
}
