% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{fetch_subscription}
\alias{fetch_subscription}
\title{Fetch Subscription}
\usage{
fetch_subscription(authorization, id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Fetch Subscription
}
