% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{fetch_transaction}
\alias{fetch_transaction}
\title{Fetch Transaction}
\usage{
fetch_transaction(authorization, id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Fetch Transaction
}
