% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{fetch_transfer}
\alias{fetch_transfer}
\title{Fetch Transfers}
\usage{
fetch_transfer(authorization, ...)

fetch_transfer(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{recipient_id}{}

\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}
}
\value{



}
\description{
Fetch Transfers

Finalize Transfers
}
