% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{fetch_transfer}
\alias{fetch_transfer}
\title{Fetch Transfers}
\usage{
fetch_transfer(authorization, transfer_id)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{transfer_id}{string. REQUIRED
An ID or code for the transfer whose details you want to retrieve.}
}
\value{

}
\description{
Fetch Transfers
}
