% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{finalize_invoices}
\alias{finalize_invoices}
\title{Finalize a Draft Invoice}
\usage{
finalize_invoices(authorization, invoice_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Finalize a Draft Invoice
}
