% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkcharge.R
\name{initiate_bulk_charge}
\alias{initiate_bulk_charge}
\title{Initiate Bulk Charge}
\usage{
initiate_bulk_charge(authorization, reference = "", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Initiate Bulk Charge
}
