% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{initiate_bulk_transfer}
\alias{initiate_bulk_transfer}
\title{Initiate Bulk Transfer}
\usage{
initiate_bulk_transfer(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Initiate Bulk Transfer
}
