% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{initiate_transfer}
\alias{initiate_transfer}
\title{Initiate Transfer}
\usage{
initiate_transfer(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{amount}{int32. REQUIRED
Amount to transfer in kobo}

\item{recipient}{string. REQUIRED
Code for transfer recipient}

\item{source}{string. REQUIRED
Where should we transfer from? Only balance for now}

\item{currency}{string}

\item{reason}{string}

\item{reference}{string
If specified, the field should be a unique identifier (in lowercase) for the object. Only -,_ and alphanumeric characters allowed.}
}
\value{

}
\description{
Status of transfer object returned will be ‘pending’ if OTP is disabled. In the event that an OTP is required, status will read ‘otp’.
}
