% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\name{list_banks}
\alias{list_banks}
\title{List Banks}
\usage{
list_banks(...)
}
\arguments{
\item{...}{Body Params}

\item{perPage}{int32.
Specify how many records you want to retrieve per page}

\item{page}{int32.
Specify exactly what page you want to retrieve}
}
\value{

}
\description{
List Banks
}
