% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{list_plans}
\alias{list_plans}
\title{List Plans}
\usage{
list_plans(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
List Plans
}
