% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{list_subscriptions}
\alias{list_subscriptions}
\title{List Subscriptions}
\usage{
list_subscriptions(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
List Subscriptions
}
