% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{list_subscriptions}
\alias{list_subscriptions}
\title{List Subscriptions}
\usage{
list_subscriptions(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
List Subscriptions
}
