% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{list_transfer_recipients}
\alias{list_transfer_recipients}
\title{List Transfer Recipients}
\usage{
list_transfer_recipients(authorization, ...)

list_transfer_recipients(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{perPage}{int32.
Specify how many records you want to retrieve per page}

\item{page}{int32.
Specify exactly what page you want to retrieve}

\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{perPage}{int32.
Specify how many records you want to retrieve per page}

\item{page}{int32.
Specify exactly what page you want to retrieve}
}
\value{



}
\description{
List Transfer Recipients

List Transfers
}
