% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{request_reauthorization}
\alias{request_reauthorization}
\title{Request Reauthorization}
\usage{
request_reauthorization(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Request Reauthorization
}
