% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\name{resolve_bvn}
\alias{resolve_bvn}
\title{Resolve BVN}
\usage{
resolve_bvn(authorization, bvn)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{bvn}{}
}
\value{

}
\description{
Resolve BVN
}
