% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_phone.R
\name{resolve_phone_number}
\alias{resolve_phone_number}
\title{Resolve Phone Number}
\usage{
resolve_phone_number(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Resolve Phone Number
}
