% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkcharge.R
\name{resume_bulk_charge_batch}
\alias{resume_bulk_charge_batch}
\title{Resume Bulk Charge Batch}
\usage{
resume_bulk_charge_batch(authorization, batch_code)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{batch_code}{}
}
\value{

}
\description{
Resume Bulk Charge Batch
}
