% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{set_risk_action}
\alias{set_risk_action}
\title{Whitelist / Blacklist Customer}
\usage{
set_risk_action(authorization, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Whitelist / Blacklist Customer
}
