% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge.R
\name{submit_birthday}
\alias{submit_birthday}
\title{Submit Birthday}
\usage{
submit_birthday(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{birthday}{date. REQUIRED Birthday submitted by user}

\item{reference}{string. REQUIRED reference for ongoing transaction}
}
\value{

}
\description{
Submit birthday when requested
}
