% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge.R
\name{submit_phone}
\alias{submit_phone}
\title{Submit Phone}
\usage{
submit_phone(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Submit Phone
}
