% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge.R
\name{submit_pin}
\alias{submit_pin}
\title{Submit Pin}
\usage{
submit_pin(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Submit Pin
}
