% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction_totals}
\alias{transaction_totals}
\title{Transaction Totals}
\usage{
transaction_totals(authorization, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}

\item{from}{date.
Lower bound of date range. Leave undefined to show totals from day one.}

\item{to}{date.
Upper bound of date range. Leave undefined to show totals till date.}
}
\value{

}
\description{
Transaction Totals
}
