% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{update_page}
\alias{update_page}
\title{Update Page}
\usage{
update_page(authorization, page_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Update Page
}
