% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payment_sessions.R
\name{update_payment_session_timeout}
\alias{update_payment_session_timeout}
\title{Update Payment Session Timeout}
\usage{
update_payment_session_timeout(authorization)
}
\arguments{
\item{authorization}{}
}
\value{

}
\description{
Update Payment Session Timeout
}
