% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subaccount.R
\name{update_subaccount}
\alias{update_subaccount}
\title{Update Subaccount}
\usage{
update_subaccount(authorization, subacct_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{...}{Body Params}
}
\value{

}
\description{
Update Subaccount
}
