% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{update_transfer_recipients}
\alias{update_transfer_recipients}
\title{Update Transfer Recipients}
\usage{
update_transfer_recipients(authorization, recipient_id, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Update Transfer Recipients
}
