% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{update_transfer_recipients}
\alias{update_transfer_recipients}
\title{Update Transfer Recipients}
\usage{
update_transfer_recipients(authorization, recipient_id, ...)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{recipient_id}{string. REQUIRED
Code or ID for transfer recipient we want to update.}

\item{...}{Body Params}

\item{name}{string. REQUIRED
A name for the recipient}

\item{email}{string.
The email address of the recipient}
}
\value{

}
\description{
Updates an existing recipient
}
