% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{verify_transaction}
\alias{verify_transaction}
\title{Verify Transaction}
\usage{
verify_transaction(authorization, reference)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{reference}{string. REQUIRED Path Param.
Unique transaction reference that exists in your account}
}
\value{

}
\description{
Verify Transaction
}
