% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{view_transaction_timeline}
\alias{view_transaction_timeline}
\title{View Transaction Timeline}
\usage{
view_transaction_timeline(authorization, id)
}
\arguments{
\item{authorization}{set_keys("", "SECRET_KEY")$secret,
equivalent of "-H Authorization: Bearer SECRET_kEY"}

\item{id}{int32. REQUIRED Path Param.
An ID for the transaction to fetch}
}
\value{

}
\description{
View Transaction Timeline
}
